requirements.txt
absl-py==0.7.1
asn1crypto==0.24.0
astor==0.7.1
attrs==19.1.0
Automat==0.7.0
bls==0.3.0
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
constantly==15.1.0
cryptography==2.6.1
cssselect==1.0.3
cycler==0.10.0
decorator==4.4.0
epa==1.3
gast==0.2.2
grpcio==1.20.1
h5py==2.9.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
joblib==0.13.2
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.1.0
lxml==4.3.3
Markdown==3.1
matplotlib==3.0.3
mock==3.0.5
mysql-connector==2.2.9
mysql-connector-python==8.0.15
numpy==1.16.3
pandas==0.24.2
parsel==1.5.1
protobuf==3.7.1
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycparser==2.19
PyDispatcher==2.0.5
PyHamcrest==1.9.0
pyOpenSSL==19.0.0
pyparsing==2.4.0
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1
pyzillow==0.5.5
queuelib==1.5.0
requests==2.21.0
scikit-learn==0.21.0
scipy==1.2.1
Scrapy==1.6.0
selenium==3.141.0
service-identity==18.1.0
six==1.12.0
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
Theano==1.0.4
Twisted==19.2.0
urllib3==1.24.3
w3lib==1.20.0
wbdata==0.2.7
Werkzeug==0.15.2
zope.interface==4.6.0
